refactor: make Functors.jl a weakdep
