+++ src/DiffEqCallbacks.jl
@@ -4,7 +4,6 @@ using ConcreteStructs: @concrete
 using DataStructures: DataStructures, BinaryMaxHeap, BinaryMinHeap
 using DiffEqBase: DiffEqBase, get_tstops, get_tstops_array, get_tstops_max
 using DifferentiationInterface: DifferentiationInterface, Constant
-using Functors: Functors, fmap
 using LinearAlgebra: LinearAlgebra, adjoint, axpy!, copyto!, mul!, ldiv!
 using Markdown: @doc_str
 using RecipesBase: @recipe
