+++ src/functor_helpers.jl
@@ -1,56 +1,36 @@
-# NOTE: `fmap` can handle all these cases without us defining them, but it often makes the
-#       code type unstable. So we define them here to make the code type stable.
-# Handle Non-Array Parameters in a Generic Fashion
 """
     recursive_copyto!(y, x)
 
 `y[:] .= vec(x)` for generic `x` and `y`. This is used to handle non-array parameters!
 """
-recursive_copyto!(y, x) = fmap(internal_copyto!, y, x; exclude = isleaf)
-
-function internal_copyto!(y, x)
-    hasmethod(copyto!, Tuple{typeof(y), typeof(x)}) ? copyto!(y, x) : nothing
-end
+function recursive_copyto! end
 
 """
     neg!(x)
 
 `x .*= -1` for generic `x`. This is used to handle non-array parameters!
 """
-recursive_neg!(x) = fmap(internal_neg!, x; exclude = isleaf)
-
-internal_neg!(x::AbstractArray) = x .*= -1
-internal_neg!(x) = nothing
-
+function recursive_neg! end
 """
     zero!(x)
 
 `x .= 0` for generic `x`. This is used to handle non-array parameters!
 """
-recursive_zero!(x) = fmap(internal_zero!, x; exclude = isleaf)
-
-internal_zero!(x::AbstractArray) = fill!(x, false)
-internal_zero!(x) = nothing
+function recursive_zero! end
 
 """
     recursive_sub!(y, x)
 
 `y .-= x` for generic `x` and `y`. This is used to handle non-array parameters!
 """
-recursive_sub!(y, x) = fmap(internal_sub!, y, x; exclude = isleaf)
-
-internal_sub!(y::AbstractArray, x::AbstractArray) = y .-= x
-internal_sub!(y, x) = nothing
+function recursive_sub! end
 
 """
     recursive_add!(y, x)
 
 `y .+= x` for generic `x` and `y`. This is used to handle non-array parameters!
 """
-recursive_add!(y, x) = fmap(internal_add!, y, x; exclude = isleaf)
-
-internal_add!(y::AbstractArray, x::AbstractArray) = y .+= x
-internal_add!(y, x) = nothing
+function recursive_add! end
 
 """
     allocate_vjp(λ, x)
@@ -58,58 +38,29 @@ internal_add!(y, x) = nothing
 
 `similar(λ, size(x))` for generic `x`. This is used to handle non-array parameters!
 """
-function allocate_vjp(λ::AbstractArray, x)
-    fmap(
-        Base.Fix1(allocate_vjp_internal, λ), x; exclude = isleaf)
-end
-allocate_vjp(x) = fmap(similar, x)
-
-allocate_vjp_internal(λ::AbstractArray, x) = similar(λ, size(x))
+function allocate_vjp end
 
 """
     allocate_zeros(x)
 
 `zero.(x)` for generic `x`. This is used to handle non-array parameters!
 """
-allocate_zeros(x) = fmap(internal_allocate_zeros, x; exclude = isleaf)
-
-internal_allocate_zeros(x) = hasmethod(zero, Tuple{typeof(x)}) ? zero(x) : nothing
+function allocate_zeros end
 
 """
 recursive_copy(y)
 
 `copy(y)` for generic `y`. This is used to handle non-array parameters!
 """
-recursive_copy(y) = fmap(internal_copy, y; exclude = isleaf)
-
-internal_copy(x) = hasmethod(copy, Tuple{typeof(x)}) ? copy(x) : nothing
+function recursive_copy end
 
 """
     recursive_adjoint(y)
 
 `adjoint(y)` for generic `y`. This is used to handle non-array parameters!
 """
-recursive_adjoint(y) = fmap(internal_adjoint, y; exclude = isleaf)
+function recursive_adjoint end
 
-internal_adjoint(x) = hasmethod(adjoint, Tuple{typeof(x)}) ? adjoint(x) : nothing
+function recursive_scalar_mul! end
 
-# scalar_mul!
-recursive_scalar_mul!(x, α) = fmap(Base.Fix2(internal_scalar_mul!, α), x; exclude = isleaf)
-
-internal_scalar_mul!(x::Number, α) = x * α
-internal_scalar_mul!(x::AbstractArray, α) = x .*= α
-internal_scalar_mul!(x, α) = nothing
-
-# axpy!
-function recursive_axpy!(α, x, y)
-    fmap((xᵢ, yᵢ) -> internal_axpy!(α, xᵢ, yᵢ), x, y; exclude = isleaf)
-end
-
-internal_axpy!(α, x::AbstractArray, y::AbstractArray) = axpy!(α, x, y)
-internal_axpy!(α, x, y) = nothing
-
-# isleaf
-isleaf(x) = Functors.isleaf(x)
-
-## BigFloat and such are not bitstype
-isleaf(::AbstractArray{T}) where {T} = isbitstype(T) || T <: Number
+function recursive_axpy! end
